refactor: extract FastUSDCConfigShape
