+++ packages/builders/scripts/fast-usdc/start-fast-usdc.build.js
@@ -1,6 +1,8 @@
 // @ts-check
 import { makeHelpers } from '@agoric/deploy-script-support';
 import { AmountMath } from '@agoric/ertp';
+import { getManifestForFastUSDC } from '@agoric/fast-usdc/src/start-fast-usdc.core.js';
+import { FastUSDCConfigShape } from '@agoric/fast-usdc/src/type-guards.js';
 import { toExternalConfig } from '@agoric/fast-usdc/src/utils/config-marshal.js';
 import { configurations } from '@agoric/fast-usdc/src/utils/deploy-config.js';
 import {
@@ -9,10 +11,6 @@ import {
 } from '@agoric/zoe/src/contractSupport/ratio.js';
 import { Far } from '@endo/far';
 import { parseArgs } from 'node:util';
-import {
-  FastUSDCConfigShape,
-  getManifestForFastUSDC,
-} from '@agoric/fast-usdc/src/start-fast-usdc.core.js';
 
 /**
  * @import {CoreEvalBuilder, DeployScriptFunction} from '@agoric/deploy-script-support/src/externalTypes.js'
